feat: Change structure for use locale route folder
