+++ app/[locale]/register/page.tsx
@@ -7,7 +7,7 @@ export default async function RegisterPage() {
   const session = await getServerSession();
 
   if (session) {
-    redirect("/");
+    redirect("/es");
   }
 
   return (
