+++ middleware.ts
@@ -2,5 +2,5 @@ export { default } from "next-auth/middleware";
 
 export const config = {
   // specify the route you want to protect
-  matcher: ["/"],
+  matcher: ["/es"],
 };
